fix: extract GalaxyView camera hook to meet 300-line file limit
Move animation loop, camera interpolation, and star/principle position
updates into useGalaxyCamera.js — brings GalaxyView.jsx from 321 to 154
lines.
